move testbed activity components to src/components
